Split MAX211000 driver headers into a register defintion file and a class defintnion file. Added max21100 to modules options
